My contribution to the alg
I'm committing the local changes I made to this project back in 2019,
but I forgot why I was even interested in this algorithm, and what
my changes are supposed to improve.

Maybe in a future life I will rekindle my interest in this and do
something amazing with it.

AS IS.